feat(pos): move KOT reprint button from order cart to right sidebar options menu
