+++ pos/src/components/OrderPanel.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { Trash2, Edit, FrownIcon, Plus, Loader2, MessageSquare, Printer } from 'lucide-react';
+import { Trash2, Edit, FrownIcon, Plus, Loader2, MessageSquare } from 'lucide-react';
 import { usePOSStore } from '../store/pos-store';
 import { cn } from '@ury/ui';
 import { formatCurrency } from '@ury/core';
@@ -9,7 +9,7 @@ import OrderTypeSelect from './OrderTypeSelect';
 import CommentDialog from './CommentDialog';
 import { Button } from '@ury/ui';
 import { Spinner } from '@ury/ui';
-import { syncOrder, reprintKOT } from '../lib/order-api';
+import { syncOrder } from '../lib/order-api';
 import { useRootStore } from '../store/root-store';
 import type { RootState } from '../store/root-store';
 import { showToast } from '@ury/ui';
@@ -42,7 +42,6 @@ const OrderPanel = () => {
   const [editingItem, setEditingItem] = useState<typeof activeOrders[0] | null>(null);
   const [isSubmitting, setIsSubmitting] = useState(false);
   const [showCommentDialog, setShowCommentDialog] = useState(false);
-  const [isReprintingKOT, setIsReprintingKOT] = useState(false);
 
   const calculateItemTotal = (item: typeof activeOrders[0]) => {
     const basePrice = item.selectedVariant?.price || item.price;
@@ -67,36 +66,6 @@ const OrderPanel = () => {
 
   const handleCommentSave = (comment: string) => {
     setOrderComment(comment);
-  };
-
-  const handleKOTReprint = async () => {
-    if (!orderId) return;
-    try {
-      setIsReprintingKOT(true);
-      const res = await reprintKOT(orderId);
-      const message = typeof res === 'string' ? res : res?.message;
-      if (typeof message === 'string' && message.startsWith('Failure')) {
-        showToast.error(message);
-      } else {
-        showToast.success(t('success.kot_reprinted') || 'KOT Reprinted successfully');
-      }
-    } catch (error: any) {
-      console.error('Failed to reprint KOT:', error);
-      // Frappe API error handling
-      if (error && typeof error === 'object' && '_server_messages' in error && typeof error._server_messages === 'string') {
-        try {
-          const messages = JSON.parse(error._server_messages);
-          const messageObj = JSON.parse(messages[0]);
-          showToast.error(messageObj.message || 'API error');
-        } catch {
-          showToast.error('API error');
-        }
-      } else {
-        showToast.error(error.message || 'Failed to reprint KOT');
-      }
-    } finally {
-      setIsReprintingKOT(false);
-    }
   };
 
   const handleSubmit = async () => {
@@ -331,18 +300,6 @@ const OrderPanel = () => {
                 >
                   <MessageSquare className="w-4 h-4" />
                 </Button>
-                {posProfile?.enable_kot_reprint === 1 && isUpdatingOrder && orderId && (
-                  <Button
-                    onClick={handleKOTReprint}
-                    variant="ghost"
-                    size="sm"
-                    className="h-8 w-8 p-0 text-gray-500 hover:text-gray-700"
-                    disabled={isInteractionDisabled || isReprintingKOT}
-                    title={t('cart.reprint_kot') || 'Reprint KOT'}
-                  >
-                    {isReprintingKOT ? <Loader2 className="w-4 h-4 animate-spin" /> : <Printer className="w-4 h-4" />}
-                  </Button>
-                )}
                 <span className="text-lg font-semibold">{t('cart.total')}</span>
               </div>
               <span className="text-lg font-semibold">{formatCurrency(total)}</span>
